Removed includeAll.pl; it's no longer needed
This script was used to create a single perl file including all subroutines.
Although this worked at the time, it really wasn't an elegant or platform
independent solution.  Since this project is being updated to more modern
programming standards, the file is no longer necessary and can be safely
deleted.