Using warnings and strict in printSearchResults.pl
printSearchResults() takes the number of entries found in the grep of the
previous subroutine (yes, this code needs to be refactored...) as its
argument.  This argument is passed in from entriesFoundDecide().

In entriesFoundDecide() the $ind variable was localised as it isn't actually
a global variable.  The variable is now passed to entriesFoundDecide as an
argument.