Returning data from printToEdit
This passes some information on from printAllFields() to other routines
which require @fieldArray and @paper data.  These variables were global and
thus now need to be passed through the subs to get them to the location
where they are needed.  Although this isn't great, it's better than using
global vars, and it can hopefully be more easily cleaned up later.